requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-xdist==3.5.0
factory-boy==3.2.1

# Code quality
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.3.0

# Documentation
sphinx==6.2.1
sphinx-rtd-theme==1.2.0

# Development server
python-dotenv==1.0.0

# Debugging
django-debug-toolbar==4.1.0
ipython==8.15.0

# Performance Profiling
pyinstrument==4.5.4
memory-profiler==0.61.0
psutil==5.9.5

# Database Tools
pgcli==4.0.1
mysqlclient==2.1.1

# Security
safety==2.3.5
bandit==1.7.4

# Pre-commit hooks
pre-commit==3.3.3

# Monitoring and Metrics
sentry-sdk==1.28.1
prometheus-client==0.17.1

# Load Testing
locust==2.15.1

# Code Analysis
pylint==3.0.2
radon==6.0.1

# Development Utilities
watchdog==3.0.0
python-dotenv==1.0.0

# Database Optimization
django-perf-rec==4.17.0
django-silk==5.1.0

# API Development
drf-yasg==1.21.5
django-extensions==3.2.3
